pyproject.toml
[build-system]
requires = ["setuptools >= 61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "discrete1"
version = "0.1.0"
dependencies = [
    "numpy",
    "numba",
    "pytest",
    "tqdm",
]
authors = [
    {name = "Ben Whewell", email = "ben.whewell@pm.me"}
]
description = "1D Discrete Ordinates Multigroup Neutron Transport"
readme = "README.md"
license = "BSD-3-Clause"
keywords = ["neutron transport", "machine learning", "DJINN"]

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra -q --cov=discrete1 --cov-report=html --cov-report=xml"
testpaths = ["tests"]

[project.optional-dependencies]
ml = [
    "scikit-learn>=1.4.0",
    "tensorflow>=2.4.0",
    "djinnml @ git+https://git@github.com/bwhewe-13/DJINN.git",
]
dev = [
    "pytest>=6.0",
    "pytest-cov",
    "black",
    "isort",
    "flake8",
    "pre-commit",
]
all = [
    "scikit-learn>=1.4.0",
    "tensorflow>=2.4.0",
    "djinnml @ git+https://git@github.com/bwhewe-13/DJINN.git",
    "pytest>=6.0",
    "pytest-cov",
    "black",
    "isort",
    "flake8",
    "pre-commit",
]
